+++ apiServices/deleteOneProduct/deleteOneProduct.js
@@ -4,6 +4,12 @@ const prisma = new PrismaClient()
 
 const deleteOneProduct = async (req, res) => {
   const productId = req.params.productID
+  
+  const sessionId = req.cookies.sessionId 
+
+  if (!sessionId) {
+    return res.status(401).json({ error: 'Unauthorized' })
+  }
 
   try {
     const deletedProduct = await prisma.products.delete({
